Minor fix in clean.sh

+++ clean.sh
@@ -3,7 +3,7 @@
 # Makefile?! We don't need no stinkin' Makefile!
 
 while IFS= read -r prog; do
-    rm -f $prog
+    [[ -x $prog ]] && rm -f $prog
 done < .gitignore
 
 find . -name screenshot.png -delete
